Moved duplicate moving code to Moving.cs :racehorse:
