+++ Assets/Scripts/Moving.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 04424a181ee824d4085bd86950775146
+timeCreated: 1471625543
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
